Refactor manangement commands tests
